chore(utils): renaming tex utils
