+++ src/dmd_toolbox/utils/__init__.py
@@ -1,4 +1,4 @@
-from .tex_utils import checkdep_usetex
+from .dependency import checkdep_usetex
 
 __all__ = [
     "checkdep_usetex"
